+++ CHANGELOG.md
@@ -7,6 +7,15 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
 ### Added
 
+### Removed
+
+### Changed
+
+
+## [0.3.0] - 2019-02-25
+
+### Added
+
 - Introduced `All` and `Any` predicates, implementing `Filter` in
   terms of the former
 - A new `ApplyOption` called `Replace` that defaults to false. Can be
@@ -115,7 +124,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
 [controller-runtime]: https://github.com/manifestival/controller-runtime-client
 [client-go]: https://github.com/manifestival/client-go-client
-[unreleased]: https://github.com/manifestival/manifestival/compare/v0.2.0...HEAD
+[unreleased]: https://github.com/manifestival/manifestival/compare/v0.3.0...HEAD
+[0.3.0]: https://github.com/manifestival/manifestival/compare/v0.2.0...v0.3.0
 [0.2.0]: https://github.com/manifestival/manifestival/compare/v0.1.0...v0.2.0
 [0.1.0]: https://github.com/manifestival/manifestival/compare/v0.0.0...v0.1.0
 [0.0.0]: https://github.com/manifestival/manifestival/releases/tag/v0.0.0